add tests for level filters
